Use X11 directly to track and change active windows.
Active windows is detected by monitoring changes on the _NET_ACTIVE_WINDOW
property on the root window.

The `switch` command is sent through a ClientMessage to the root window.

The program no longer tries to keep a queue of known windows. Instead, it just
track the last visible window.

If another window is activated before releasing the keyboard modifiers, the
reference to the `last` window is kept.